# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Toko API",
        "description": "API documentation for Toko API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/bill/{table_number}": {
            "get": {
                "tags": [
                    "Bill"
                ],
                "summary": "Get the bill for a specific table",
                "operationId": "e73c369355c2d50cb17aa0abb7bc00b5",
                "parameters": [
                    {
                        "name": "table_number",
                        "in": "path",
                        "description": "The table number",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "MEJA NO 1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bill retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "order_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "table_number": {
                                            "type": "string",
                                            "example": "MEJA NO 1"
                                        },
                                        "total_price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 89000
                                        },
                                        "bill_details": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "product_name": {
                                                        "type": "string",
                                                        "example": "Jeruk"
                                                    },
                                                    "variant": {
                                                        "type": "string",
                                                        "example": "DINGIN"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 12000
                                                    },
                                                    "quantity": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "subtotal": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 12000
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Table not found or no orders found for this table"
                    }
                }
            }
        },
        "/api/order": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Create a new order",
                "operationId": "34d0f49fe3fcbbc0ce90006ce29412df",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "table_number": {
                                        "type": "string",
                                        "example": "MEJA NO 1"
                                    },
                                    "order_items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "order_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total_price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 89000
                                        },
                                        "printers": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Bar"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Table not found"
                    },
                    "500": {
                        "description": "Order processing failed"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Bill",
            "description": "Operations related to bills"
        },
        {
            "name": "Order",
            "description": "Order"
        }
    ]
}